fix: Resolve MSVC signed/unsigned and narrowing warnings across codebase
- Cast -1 sentinel values to static_cast<uint64_t>(-1) in ELF/PE/MachO/Strategy handlers
- Fix int-to-uint16_t narrowing in opcode_table OID computation
- Fix size_t-to-uint16_t truncation in instruction_t hash return
- Fix uint16_t-to-uint8_t narrowing in file_type_parser PE mode lookup
- Use explicit uint16_t keys in ELF/PE arch lookup tables
- Rename shadowed variable in ELFHandler::getEntryIndex

+++ sdk/src/segmentator/MachOHandler.cpp
@@ -46,7 +46,7 @@ MachOFileHandlerStrategy::~MachOFileHandlerStrategy() = default;
 std::pair<uint64_t, uint64_t>
 MachOFileHandlerStrategy::doGetBeginEndAddr() noexcept {
     // Find stub addresses for VMPilot_Begin and VMPilot_End
-    uint64_t begin_addr = -1, end_addr = -1;
+    uint64_t begin_addr = static_cast<uint64_t>(-1), end_addr = static_cast<uint64_t>(-1);
     for (const auto& entry : pImpl->parser.stubEntries()) {
         auto name = stripLeadingUnderscore(entry.symbol_name);
         if (name == VMPilot::Common::BEGIN_VMPILOT_SIGNATURE) {
@@ -72,7 +72,7 @@ uint64_t MachOFileHandlerStrategy::doGetTextBaseAddr() noexcept {
         auto sect = pImpl->parser.findSection("__TEXT", "__text");
         if (!sect) {
             spdlog::error("Could not find __TEXT,__text section");
-            return -1;
+            return static_cast<uint64_t>(-1);
         }
         pImpl->text_base_addr = sect->addr;
     }
